Rename lodash to underscore. It's just the NAME of a MOCK object, anyway. ;)
