+++ tests/js/unit/stubs/underscore.js
